update jwt related code
- moved 'authorize' decorator to src/auth/utils
- added group-admin enforcement to create_group_type endpoint
- added test for the protected endpoint
